+++ packages/webview_windows/windows/webview_host.h
@@ -42,8 +42,8 @@ class WebviewHost {
       std::optional<std::wstring> browser_exe_path = std::nullopt,
       std::optional<std::string> arguments = std::nullopt);
 
-  void CreateWebview(HWND hwnd, bool offscreen_only, bool owns_window,
-                     WebviewCreationCallback callback);
+  void CreateWebview(HWND hwnd, HWND flutter_view_hwnd, bool offscreen_only,
+                     bool owns_window, WebviewCreationCallback callback);
 
   void CreateWebViewPointerInfo(PointerInfoCreationCallback cb);